+++ lib/widgets/expats_wiki_button.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:portfolio/core/utils/constants.dart';
-import 'package:portfolio/core/utils/navigate_links.dart';
+import 'package:portfolio/shared/constants.dart';
+import 'package:portfolio/shared/navigate_links.dart';
 
 class ExpatsWikiButton extends StatelessWidget {
   const ExpatsWikiButton({super.key});
